With great power comes great responsibility
The Africana Framework is software designed for network and web hacking, automating as many tasks as possible to detect vulnerabilities in common services and web technologies. It supports a wide range of penetration testing, including internal network assessments, Wi-Fi security, system anonymity, and web bug hunting. It is purely written for Good and not Evil.